Refactor the reporter and add an YAML report format
Separate the events collection from report rendering, in order
to allow different type of reports.
