+++ kube-hunter.py
@@ -3,8 +3,6 @@ from __future__ import print_function
 
 import argparse
 import logging
-import sys
-import time
 
 try:
     raw_input          # Python 2
@@ -20,6 +18,7 @@ parser.add_argument('--mapping', action="store_true", help="outputs only a mappi
 parser.add_argument('--remote', nargs='+', metavar="HOST", default=list(), help="one or more remote ip/dns to hunt")
 parser.add_argument('--active', action="store_true", help="enables active hunting")
 parser.add_argument('--log', type=str, metavar="LOGLEVEL", default='INFO', help="set log level, options are: debug, info, warn, none")
+parser.add_argument('--report', type=str, default='plain', help="set report type, options are: plain, yaml")
 
 import plugins
 
@@ -31,6 +30,14 @@ except:
     pass
 if config.log.lower() != "none":
     logging.basicConfig(level=loglevel, format='%(message)s', datefmt='%H:%M:%S')
+
+from src.modules.report.plain import PlainReporter
+from src.modules.report.yaml import YAMLReporter
+
+if config.report.lower() == "yaml":
+    config.reporter = YAMLReporter()
+else:
+    config.reporter = PlainReporter()
 
 from src.core.events import handler
 from src.core.events.types import HuntFinished, HuntStarted
