+++ requirements.txt
@@ -5,3 +5,4 @@ scapy
 requests
 PrettyTable
 urllib3
+ruamel.yaml
